// src/interfaces/empresa.ts
export interface Empresa {
    tipo: string;
    nome: string;
    uf: string;
    telefone: string;
    situacao: string;
    bairro: string;
    logradouro: string;
    numero: number;
    cep: string;
    municipio: string;
    porte: string;
    abertura: string;
    natureza_juridica: string;
    fantasia: string;
    cnpj: string;
    status: string;
    complemento: string;
    email: string;
    message: string;
}